Update chipseq-se not to fail on MACS2
When MACS2 fails to build model, it
stops with error. To avoid it --fix-bimodal
is used. It means that if we run MACS2 with
--extsize
--bw
--fix-bimodal
and --nomodel is not set, MASCS2 will try
to build shifting model and in a case of
failure will use --extsize to run it as
if --nomodel had not been set.
If we set --nomodel at the very beginning
--extsize isnot used at all, only --bw is
used.