Add ATAC-Seq SE/PE pipelines
Based on the ChIP-Seq SE/PE with Trimgalore.
The only difference is that we set fragment size
to be 9 bp on the genome coverage step generation.
The latter will allow to use only the first 9 bp
from every read. Addiionaly, for PE we have to
unset option 'pairchip'. Otherwise, fragment size
set to 9 bp is ignored